requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
discord.py==2.2.3
docopt==0.6.2
Flask==2.2.2
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgspec==0.9.1
multidict==6.0.4
pdf2image==1.16.3
Pillow==9.3.0
pipreqs==0.4.11
pyzbar==0.1.9
requests==2.28.1
urllib3==1.26.13
valo-api==1.5.4
Werkzeug==2.2.2
yarg==0.1.9
yarl==1.9.2
